+++ include/databento/file_stream.hpp
@@ -2,16 +2,16 @@
 
 #include <cstddef>  // size_t
 #include <cstdint>  // uint8_t
-#include <fstream>  // ifstream
+#include <fstream>  // ifstream, ofstream
 #include <string>
 
 #include "databento/ireadable.hpp"
+#include "databento/iwritable.hpp"
 
 namespace databento {
-namespace detail {
-class FileStream : public IReadable {
+class InFileStream : public IReadable {
  public:
-  explicit FileStream(const std::string& file_path);
+  explicit InFileStream(const std::string& file_path);
 
   // Read exactly `length` bytes into `buffer`.
   void ReadExact(std::uint8_t* buffer, std::size_t length) override;
@@ -22,5 +22,14 @@ class FileStream : public IReadable {
  private:
   std::ifstream stream_;
 };
-}  // namespace detail
+
+class OutFileStream : public IWritable {
+ public:
+  explicit OutFileStream(const std::string& file_path);
+
+  void WriteAll(const std::uint8_t* buffer, std::size_t length) override;
+
+ private:
+  std::ofstream stream_;
+};
 }  // namespace databento