+++ src/dbn_file_store.cpp
@@ -3,27 +3,30 @@
 #include <memory>   // unique_ptr
 #include <utility>  // move
 
-#include "databento/detail/file_stream.hpp"
+#include "databento/file_stream.hpp"
 #include "databento/ireadable.hpp"
+#include "databento/record.hpp"
 
 using databento::DbnFileStore;
 
 DbnFileStore::DbnFileStore(const std::string& file_path)
-    : parser_{detail::FileStream{file_path}} {}
+    : decoder_{ILogReceiver::Default(), InFileStream{file_path}} {}
 
-DbnFileStore::DbnFileStore(const std::string& file_path,
+DbnFileStore::DbnFileStore(ILogReceiver* log_receiver,
+                           const std::string& file_path,
                            VersionUpgradePolicy upgrade_policy)
-    : parser_{std::unique_ptr<IReadable>{new detail::FileStream{file_path}},
-              upgrade_policy} {}
+    : decoder_{log_receiver,
+               std::unique_ptr<IReadable>{new InFileStream{file_path}},
+               upgrade_policy} {}
 
 void DbnFileStore::Replay(const MetadataCallback& metadata_callback,
                           const RecordCallback& record_callback) {
-  auto metadata = parser_.DecodeMetadata();
+  auto metadata = decoder_.DecodeMetadata();
   if (metadata_callback) {
     metadata_callback(std::move(metadata));
   }
   const databento::Record* record;
-  while ((record = parser_.DecodeRecord()) != nullptr) {
+  while ((record = decoder_.DecodeRecord()) != nullptr) {
     if (record_callback(*record) == KeepGoing::Stop) {
       break;
     }
@@ -32,4 +35,21 @@ void DbnFileStore::Replay(const MetadataCallback& metadata_callback,
 
 void DbnFileStore::Replay(const RecordCallback& record_callback) {
   Replay({}, record_callback);
+}
+
+const databento::Metadata& DbnFileStore::GetMetadata() {
+  MaybeDecodeMetadata();
+  return metadata_;
+}
+
+const databento::Record* DbnFileStore::NextRecord() {
+  MaybeDecodeMetadata();
+  return decoder_.DecodeRecord();
+}
+
+void DbnFileStore::MaybeDecodeMetadata() {
+  if (!has_decoded_metadata_) {
+    metadata_ = decoder_.DecodeMetadata();
+    has_decoded_metadata_ = true;
+  }
 }